requirements.txt
requests
gradio
matplotlib
pdfminer.six
Pillow
tqdm
